Saving code for Openstack neutron refresh
Saving code for Openstack neutron refresh, both Cloud and Infra
