+++ app/models/ems_refresh/save_inventory_cloud.rb
@@ -56,7 +56,9 @@ module EmsRefresh::SaveInventoryCloud
       :cloud_volumes,
       :cloud_volume_snapshots,
       :vms,
+      :network_routers,
       :floating_ips,
+      :network_ports,
       :cloud_resource_quotas,
       :cloud_object_store_containers,
       :cloud_object_store_objects
@@ -66,6 +68,7 @@ module EmsRefresh::SaveInventoryCloud
     save_child_inventory(ems, hashes, child_keys, target)
 
     link_volumes_to_base_snapshots(hashes[:cloud_volumes]) if hashes.key?(:cloud_volumes)
+    link_floating_ips_to_network_ports(hashes[:floating_ips]) if hashes.key?(:floating_ips)
 
     ems.save!
     hashes[:id] = ems.id
@@ -147,99 +150,6 @@ module EmsRefresh::SaveInventoryCloud
 
     save_inventory_multi(:key_pairs, ems, hashes, deletes, [:name])
     store_ids_for_new_records(ems.key_pairs, hashes, :name)
-  end
-
-  def save_cloud_networks_inventory(ems, hashes, target = nil)
-    target = ems if target.nil?
-
-    ems.cloud_networks(true)
-    deletes = if (target == ems)
-                ems.cloud_networks.dup
-              else
-                []
-              end
-
-    hashes.each do |h|
-      h[:cloud_tenant_id]        = h.fetch_path(:cloud_tenant, :id)
-      h[:orchestration_stack_id] = h.fetch_path(:orchestration_stack, :id)
-    end
-
-    save_inventory_multi(:cloud_networks,
-                         ems,
-                         hashes,
-                         deletes,
-                         [:ems_ref],
-                         :cloud_subnets,
-                         [:cloud_tenant, :orchestration_stack])
-    store_ids_for_new_records(ems.cloud_networks, hashes, :ems_ref)
-  end
-
-  def save_cloud_subnets_inventory(cloud_network, hashes)
-    deletes = cloud_network.cloud_subnets(true).dup
-
-    hashes.each do |h|
-      h[:availability_zone_id] = h.fetch_path(:availability_zone, :id)
-    end
-
-    save_inventory_multi(:cloud_subnets, cloud_network, hashes, deletes, [:ems_ref], nil, :availability_zone)
-
-    cloud_network.save!
-    store_ids_for_new_records(cloud_network.cloud_subnets, hashes, :ems_ref)
-  end
-
-  def save_security_groups_inventory(ems, hashes, target = nil)
-    target = ems if target.nil?
-
-    ems.security_groups(true)
-    deletes = if (target == ems)
-                ems.security_groups.dup
-              else
-                []
-              end
-
-    hashes.each do |h|
-      h[:cloud_network_id]       = h.fetch_path(:cloud_network, :id)
-      h[:cloud_tenant_id]        = h.fetch_path(:cloud_tenant, :id)
-      h[:orchestration_stack_id] = h.fetch_path(:orchestration_stack, :id)
-    end
-
-    save_inventory_multi(:security_groups,
-                         ems, hashes,
-                         deletes,
-                         [:ems_ref],
-                         :firewall_rules,
-                         [:cloud_network, :cloud_tenant, :orchestration_stack])
-    store_ids_for_new_records(ems.security_groups, hashes, :ems_ref)
-
-    # Reset the source_security_group_id for the firewall rules after all
-    #   security groups have been saved and ids obtained.
-    firewall_rule_hashes = hashes.collect { |h| h[:firewall_rules] }.flatten.index_by { |h| h[:id] }
-    firewall_rules       = ems.security_groups.collect(&:firewall_rules).flatten
-    firewall_rules.each do |fr|
-      fr_hash = firewall_rule_hashes[fr.id]
-      fr_hash[:source_security_group_id] = fr_hash.fetch_path(:source_security_group, :id)
-      fr.update_attribute(:source_security_group_id, fr_hash[:source_security_group_id])
-    end
-  end
-
-  def save_floating_ips_inventory(ems, hashes, target = nil)
-    target = ems if target.nil?
-
-    ems.floating_ips(true)
-    deletes = if (target == ems)
-                ems.floating_ips.dup
-              else
-                []
-              end
-
-    hashes.each do |h|
-      h[:vm_id] = h.fetch_path(:vm, :id)
-      # floating ip tenants are not supported with nova network
-      h[:cloud_tenant_id] = h.fetch_path(:cloud_tenant, :id) if h.key?(:cloud_tenant)
-    end
-
-    save_inventory_multi(:floating_ips, ems, hashes, deletes, [:ems_ref], nil, [:vm, :cloud_tenant])
-    store_ids_for_new_records(ems.floating_ips, hashes, :ems_ref)
   end
 
   def save_cloud_volumes_inventory(ems, hashes, target = nil)
